requirements.txt
argparse
nibabel
numpy
pillow
tqdm
